refactor main ready to merge

+++ src/server/gamecontrol.rs
@@ -43,7 +43,7 @@ pub mod game
             }
         }
 
-        pub fn get_team_and_push(& mut self, teamname: &String, id: u128)
+        pub fn get_team_and_push(& mut self, teamname: &String, id: u32)
         {
             for i in & mut self.teams
             {
